+++ src/pkg/domain/user.go
@@ -5,7 +5,7 @@ import "fmt"
 type Users []User
 
 type User struct {
-	ID        int64
+	ID        int
 	FirstName string
 	LastName  string
 }
@@ -18,7 +18,7 @@ func (u *User) IsEmpty() bool {
 	return u.ID == 0 || u.FirstName == "" || u.LastName == ""
 }
 
-func (us *Users) FindById(id int64) User {
+func (us *Users) FindById(id int) User {
 	for _, u := range *us {
 		if u.ID == id {
 			return u
